requirements.txt
fastapi
uvicorn
requests
python-telegram-bot
sqlite3
python-dotenv